+++ src/main/java/dk/kb/datahandler/preservica/client/DsPreservicaClient.java
@@ -1,7 +1,9 @@
 package dk.kb.datahandler.preservica.client;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import dk.kb.datahandler.config.ServiceConfig;
 import dk.kb.datahandler.preservica.AccessResponseObject;
+import io.swagger.models.auth.In;
 import org.apache.http.client.utils.URIBuilder;
 import org.asynchttpclient.AsyncHttpClient;
 import org.asynchttpclient.Dsl;
@@ -15,6 +17,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.ProtocolException;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
@@ -58,6 +62,20 @@ public class DsPreservicaClient {
 
         // Call method to get first accesToken and refreshToken.
         getInitialAccess();
+    }
+
+    /**
+     * Get Preservica Client for use
+     * @return the {@link DsPreservicaClient}.
+     */
+    public static DsPreservicaClient getPreservicaClient() {
+
+        String preservicaUrl = ServiceConfig.getConfig().getString("preservica.baseUrl");
+        String user = ServiceConfig.getConfig().getString("preservica.user");
+        String password =  ServiceConfig.getConfig().getString("preservica.password");
+
+        DsPreservicaClient client = new DsPreservicaClient(preservicaUrl, user, password, 600);
+        return client;
     }
 
     /**
@@ -205,46 +223,45 @@ public class DsPreservicaClient {
         return connection;
     }
 
-    public InputStream getPreservicaObjectDetails(String id) {
-        try (AsyncHttpClient asyncHttpClient = Dsl.asyncHttpClient(Dsl.config())) {
-            StringBuilder idBuilder = new StringBuilder();
+    public InputStream getPreservicaObjectDetails(String id) throws IOException, URISyntaxException {
+        StringBuilder idBuilder = new StringBuilder();
+        URL url = new URIBuilder(baseUrl)
+                .setPathSegments(objectDetailsEndpoint)
+                // This ID needs to be prefixed with the string: sdb:IO|
+                .addParameter("id", idBuilder.append("sdb:IO|").append(id).toString())
+                .build()
+                .toURL();
 
-            URL url = new URIBuilder(baseUrl)
-                    .setPathSegments(objectDetailsEndpoint)
-                    // This ID needs to be prefixed with the string: sdb:IO|
-                    .addParameter("id", idBuilder.append("sdb:IO|").append(id).toString())
-                    .build()
-                    .toURL();
-
-        /*HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-        connection.setRequestMethod("GET");
-        connection.setRequestProperty("accept", "application/json");
-        connection.setRequestProperty("Preservica-Access-Token", accessToken);*/
+        /*try (AsyncHttpClient asyncHttpClient = Dsl.asyncHttpClient(Dsl.config())) {
 
             // Perform a GET request
-            CompletableFuture<Response> futureResponse = asyncHttpClient.prepareGet(url.toString())
+            return asyncHttpClient.prepareGet(url.toString())
                     .setHeader("accept", "application/json")
                     .setHeader("Preservica-Access-Token", accessToken)
                     .execute()
-                    .toCompletableFuture();
+                    .toCompletableFuture()
+                    .thenApply(Response::getResponseBodyAsStream)
+                    .join();   // Wait for the request to complete
 
-
-            return futureResponse.thenApply(Response::getResponseBodyAsStream).join();  // Wait for the request to complete
         } catch ( Exception e) { // TODO: Update whicxh exception this throws
             log.info("Logging: ", e);
-        }
-        return null;
+        }*/
+
+        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+
+        connection.setRequestMethod("GET");
+        connection.setRequestProperty("accept", "application/json");
+        connection.setRequestProperty("Preservica-Access-Token", accessToken);
+
+        return connection.getInputStream();
+        //return null;
     }
 
     /**
      * Interrupt the Timer, so that no more Preservica Tokens are resolved.
      */
     public void endTimer(){
-        timerThread.interrupt();
-        if (timerThread.isInterrupted()){
-            tokenTimer.cancel();
-        }
+        tokenTimer.cancel();
         log.info("Interrupted TokenTimer");
     }
 }
