+++ src/main/resources/manifestationTimes.txt
@@ -8,3 +8,14 @@
 2024.05.21.15.20.22: FINISHED MANIFESTATION PLUGIN
 2024.05.22.08.24.57: STARTED MANIFESTATION PLUGIN 22000 records in 25 minuts with profiler activated
 2024.05.22.08.59.46: FINISHED MANIFESTATION PLUGIN
+2024.05.22.10.06.26: STARTED MANIFESTATION PLUGIN
+2024.05.22.10.36.51: STARTED MANIFESTATION PLUGIN 2000 records in 5 minuts. 6.6 records pr second with AsynHttpClient and parallel streams
+2024.05.22.10.41.50: FINISHED MANIFESTATION PLUGIN
+2024.05.22.10.50.10: STARTED MANIFESTATION PLUGIN 2000 records in 3,5 minuts. 9,5 records pr second with HttpUrlConnection and parallel streams
+2024.05.22.10.53.41: FINISHED MANIFESTATION PLUGIN
+2024.05.22.10.56.42: STARTED MANIFESTATION PLUGIN
+2024.05.22.11.04.30: STARTED MANIFESTATION PLUGIN
+2024.05.22.11.12.08: STARTED MANIFESTATION PLUGIN
+2024.05.22.11.12.11: FINISHED MANIFESTATION PLUGIN
+2024.05.22.11.23.49: STARTED MANIFESTATION PLUGIN
+2024.05.22.11.26.59: STARTED MANIFESTATION PLUGIN
